siamese/requirements.txt
torch
numpy
Pillow
mtcnn
opencv-python
matplotlib
torchvision